coalesce custom-elements tests into browser suite
