+++ test/runtime-browser/browser.test.js
@@ -2,9 +2,9 @@ import { chromium } from '@playwright/test';
 import * as fs from 'fs';
 import * as path from 'path';
 import { rollup } from 'rollup';
-import { pretty_print_browser_assertion, try_load_config } from '../helpers.js';
-import * as svelte from '../../compiler.mjs';
-import { beforeAll, describe, afterAll, assert } from 'vitest';
+import { mkdirp, pretty_print_browser_assertion, try_load_config } from '../helpers.js';
+import * as svelte from '../../src/compiler';
+import { beforeAll, describe, afterAll, assert, it } from 'vitest';
 
 const internal = path.resolve('internal/index.mjs');
 const index = path.resolve('index.mjs');
@@ -12,28 +12,30 @@ const index = path.resolve('index.mjs');
 const main = fs.readFileSync(`${__dirname}/driver.js`, 'utf-8');
 const browser_assert = fs.readFileSync(`${__dirname}/assert.js`, 'utf-8');
 
+/** @type {import('@playwright/test').Browser} */
+let browser;
+
+beforeAll(async () => {
+	browser = await chromium.launch();
+	console.log('[runtime-browser] Launched browser');
+}, 20000);
+
+afterAll(async () => {
+	if (browser) await browser.close();
+});
+
 describe(
 	'runtime (browser)',
-	async (it) => {
-		/** @type {import('@playwright/test').Browser} */
-		let browser;
-
-		beforeAll(async () => {
-			browser = await chromium.launch();
-			console.log('[runtime-browser] Launched browser');
-		});
-
-		afterAll(async () => {
-			if (browser) await browser.close();
-		});
-
+	async () => {
 		const failed = new Set();
 
-		async function runTest(dir, hydrate) {
+		async function run_test(dir, hydrate) {
 			if (dir[0] === '.') return;
 
+			const cwd = `${__dirname}/samples/${dir}`;
+
 			// TODO: Vitest currently doesn't register a watcher because the import is hidden
-			const config = await try_load_config(`${__dirname}/samples/${dir}/_config.js`);
+			const config = await try_load_config(`${cwd}/_config.js`);
 			const solo = config.solo || /\.solo/.test(dir);
 			const skip = config.skip || /\.skip/.test(dir);
 
@@ -44,13 +46,14 @@ describe(
 			it_fn(`${dir} ${hydrate ? '(with hydration)' : ''}`, async () => {
 				if (failed.has(dir)) {
 					// this makes debugging easier, by only printing compiled output once
-					throw new Error('skipping test, already failed');
+					assert.fail('skipping test, already failed');
 				}
 
 				const warnings = [];
 
 				const bundle = await rollup({
 					input: 'main',
+
 					plugins: [
 						{
 							name: 'testing-runtime-browser',
@@ -85,8 +88,8 @@ describe(
 								if (id === 'main') {
 									return main.replace('__HYDRATE__', hydrate ? 'true' : 'false');
 								}
-								return null;
 							},
+
 							transform(code, id) {
 								if (id.endsWith('.svelte')) {
 									const compiled = svelte.compile(code.replace(/\r/g, ''), {
@@ -96,17 +99,10 @@ describe(
 										accessors: 'accessors' in config ? config.accessors : true
 									});
 
-									const out_dir = `${__dirname}/samples/${dir}/_output/${
-										hydrate ? 'hydratable' : 'normal'
-									}`;
+									const out_dir = `${cwd}/_output/${hydrate ? 'hydratable' : 'normal'}`;
 									const out = `${out_dir}/${path.basename(id).replace(/\.svelte$/, '.js')}`;
 
-									if (fs.existsSync(out)) {
-										fs.unlinkSync(out);
-									}
-									if (!fs.existsSync(out_dir)) {
-										fs.mkdirSync(out_dir, { recursive: true });
-									}
+									mkdirp(out_dir);
 
 									fs.writeFileSync(out, compiled.js.code, 'utf8');
 
@@ -163,9 +159,109 @@ describe(
 
 		await Promise.all(
 			fs.readdirSync(`${__dirname}/samples`).map(async (dir) => {
-				await runTest(dir, false);
-				await runTest(dir, true);
+				await run_test(dir, false);
+				await run_test(dir, true);
 			})
+		);
+	},
+	// Browser tests are brittle and slow on CI
+	{ timeout: 20000, retry: process.env.CI ? 1 : 0 }
+);
+
+describe(
+	'custom-elements',
+	async () => {
+		async function run_test(dir) {
+			if (dir[0] === '.') return;
+			const cwd = `${__dirname}/custom-elements-samples/${dir}`;
+
+			const solo = /\.solo$/.test(dir);
+			const skip = /\.skip$/.test(dir);
+
+			const warnings = [];
+			const it_fn = solo ? it.only : skip ? it.skip : it;
+
+			it_fn(dir, async () => {
+				// TODO: Vitest currently doesn't register a watcher because the import is hidden
+				const config = await try_load_config(`${cwd}/_config.js`);
+
+				const expected_warnings = config.warnings || [];
+
+				const bundle = await rollup({
+					input: `${cwd}/test.js`,
+
+					plugins: [
+						{
+							name: 'plugin-resolve-svelte',
+							resolveId(importee) {
+								if (importee === 'svelte/internal' || importee === './internal') {
+									return internal;
+								}
+
+								if (importee === 'svelte') {
+									return index;
+								}
+
+								if (importee === 'assert') {
+									return 'assert';
+								}
+							},
+
+							load(id) {
+								if (id === 'assert') return browser_assert;
+							},
+
+							transform(code, id) {
+								if (id.endsWith('.svelte')) {
+									const compiled = svelte.compile(code.replace(/\r/g, ''), {
+										customElement: true,
+										dev: config.dev
+									});
+
+									compiled.warnings.forEach((w) => warnings.push(w));
+
+									return compiled.js;
+								}
+							}
+						}
+					]
+				});
+
+				const generated_bundle = await bundle.generate({ format: 'iife', name: 'test' });
+
+				function assertWarnings() {
+					if (expected_warnings) {
+						assert.deepStrictEqual(
+							warnings.map((w) => ({
+								code: w.code,
+								message: w.message,
+								pos: w.pos,
+								start: w.start,
+								end: w.end
+							})),
+							expected_warnings
+						);
+					}
+				}
+
+				const page = await browser.newPage();
+				page.on('console', (type) => {
+					console[type.type()](type.text());
+				});
+				await page.setContent('<main></main>');
+				await page.evaluate(generated_bundle.output[0].code);
+				const test_result = await page.evaluate(`test(document.querySelector('main'))`);
+
+				if (test_result) console.log(test_result);
+
+				assertWarnings();
+
+				await page.close();
+			});
+		}
+
+		await Promise.all(
+			fs.readdirSync(`${__dirname}/custom-elements-samples`).map((dir) => run_test(dir))
 		);
 	},
 	// Browser tests are brittle and slow on CI
