+++ test/runtime-browser/custom-elements-samples/$$props/main.svelte
