+++ test/runtime-browser/custom-elements-samples/$$props/test.js
