+++ test/runtime-browser/custom-elements-samples/$$slot-dynamic-content/my-widget.svelte
