+++ test/runtime-browser/custom-elements-samples/camel-case-attribute/test.js
