+++ test/runtime-browser/custom-elements-samples/ce-options-valid/main.svelte
