+++ test/runtime-browser/custom-elements-samples/custom-method/main.svelte
