+++ test/runtime-browser/custom-elements-samples/custom-method/test.js
