+++ test/runtime-browser/custom-elements-samples/events/main.svelte
