+++ test/runtime-browser/custom-elements-samples/extended-builtin/_config.js
