+++ test/runtime-browser/custom-elements-samples/extended-builtin/custom-button.js
