+++ test/runtime-browser/custom-elements-samples/extended-builtin/main.svelte
