+++ test/runtime-browser/custom-elements-samples/nested/Counter.svelte
