+++ test/runtime-browser/custom-elements-samples/new-styled/main.svelte
