+++ test/runtime-browser/custom-elements-samples/no-missing-prop-warnings/_config.js
