+++ test/runtime-browser/custom-elements-samples/no-shadow-dom/main.svelte
