+++ test/runtime-browser/custom-elements-samples/no-tag/_config.js
