+++ test/runtime-browser/custom-elements-samples/oncreate/main.svelte
