+++ test/runtime-browser/custom-elements-samples/ondestroy/main.svelte
