+++ test/runtime-browser/custom-elements-samples/ondestroy/test.js
