+++ test/runtime-browser/custom-elements-samples/props/my-widget.svelte
