+++ test/runtime-browser/custom-elements-samples/reflect-attributes/main.svelte
