+++ test/runtime-browser/custom-elements-samples/reflect-attributes/my-widget.svelte
